chore(frontend): refactor add entry form page
