+++ apps/frontends/omni-book/app/routes/admin/add-entry.tsx
@@ -1,30 +1,28 @@
 import { useEffect, useState } from "react";
-import makeAnimated from "react-select/animated";
-import Select from "react-select";
 import type { StylesConfig } from "react-select";
 import { useFetcher } from "react-router";
 import { v4 as uuidv4 } from "uuid";
-import type { z } from "zod";
 import MediaForm from "../../components/admin/entries-config/media-form";
+import GeneralSection from "../../components/admin/entries-config/general-section";
 import type { Route } from "./+types/add-entry";
 import { checkCookie } from "~/server/utils";
 import { env } from "~/lib/env";
 import type { paths } from "~/lib/api-types";
-import { TextArea } from "~/components/ui/text-area";
 import {
   entrySchema,
   type EntryFormData,
   type EntryFormInput,
+  type FormattedEntryErrors,
 } from "~/schemas/entry-schema";
 import { Button } from "~/components/ui/button";
 import { PlusIcon } from "~/components/icons";
 import Tooltip from "~/components/ui/tooltip";
 import { Notification } from "~/components/ui/notification";
+import DescriptionSection from "~/components/admin/entries-config/description-section";
+import GenresSection from "~/components/admin/entries-config/genres-section";
 
 type Profile =
   paths["/api/profile"]["get"]["responses"]["200"]["content"]["application/json"];
-
-type FormattedEntryErrors = z.inferFormattedError<typeof entrySchema>;
 
 export const loader = async ({ request }: Route.LoaderArgs) => {
   const cookieHeader = checkCookie(request);
@@ -45,14 +43,14 @@ export const loader = async ({ request }: Route.LoaderArgs) => {
   return { user };
 };
 
-type GenreOption = {
+type SelectOption = {
   value: string;
   label: string;
 };
 
 type ArrayFormItem = EntryFormInput["media"][number] & { id: string };
 
-export const selectStyles: StylesConfig<GenreOption> = {
+export const selectStyles: StylesConfig<SelectOption> = {
   control: (styles) => ({
     ...styles,
     backgroundColor: "card-secondary",
@@ -287,8 +285,6 @@ const AddEntry = () => {
     { value: "philosophy", label: "Philosophy" },
   ];
 
-  const animatedComponents = makeAnimated();
-
   const fetcher = useFetcher();
   const [notification, setNotification] = useState(false);
 
@@ -343,246 +339,28 @@ const AddEntry = () => {
         <h1 className="text-2xl font-bold mb-4">Add Entry</h1>
         <div className="flex flex-col gap-5">
           <div className="bg-card-secondary h-[2px] w-full rounded-lg"></div>
-          <div className="flex w-full gap-4">
-            <div className="w-1/4">
-              <h1 className="text-xl font-medium">General</h1>
-              <h4 className="text-sm font-light">
-                Add general information about the book
-              </h4>
-            </div>
-            <div className="flex-1 flex ">
-              <div className="flex-1 flex flex-col gap-4">
-                <div className="flex gap-4">
-                  <label className="flex-1">
-                    <div className="flex flex-col gap-1">
-                      <p className="text-sm font-light">
-                        Title<span className="text-red-500">*</span>
-                      </p>
-                      <input
-                        name="title"
-                        type="text"
-                        className="px-4 py-2 bg-card-secondary rounded-lg outline-none placeholder:text-sm"
-                        placeholder="The Republic"
-                        required
-                        onChange={(e) => handleChange("title", e.target.value)}
-                      />
-                      {touchedFields.title && errors?.title && (
-                        <p className="text-red-500 text-xs">
-                          {errors?.title._errors[0].toString()}
-                        </p>
-                      )}
-                    </div>
-                  </label>
-                  <label className="flex-1">
-                    <div className="flex flex-col gap-1">
-                      <p className="text-sm font-light">Localized Title</p>
-                      <input
-                        name="localized-title"
-                        type="text"
-                        className="px-4 py-2 bg-card-secondary rounded-lg outline-none placeholder:text-sm"
-                        placeholder="The Republic"
-                        onChange={(e) =>
-                          handleChange("localizedTitle", e.target.value)
-                        }
-                      />
-                      {errors?.localizedTitle && (
-                        <p className="text-red-500 text-sm">
-                          {errors?.localizedTitle._errors[0].toString()}
-                        </p>
-                      )}
-                    </div>
-                  </label>
-                </div>
-                <div className="flex gap-4">
-                  <label className="flex-1">
-                    <div className="flex flex-col gap-1">
-                      <p className="text-sm font-light">
-                        Author<span className="text-red-500">*</span>
-                      </p>
-                      <input
-                        name="author"
-                        type="text"
-                        className="px-4 py-2 bg-card-secondary rounded-lg outline-none placeholder:text-sm"
-                        placeholder="Plato"
-                        required
-                        onChange={(e) => handleChange("author", e.target.value)}
-                      />
-                      {touchedFields.author && errors?.author && (
-                        <p className="text-red-500 text-xs">
-                          {errors?.author._errors[0].toString()}
-                        </p>
-                      )}
-                    </div>
-                  </label>
-                  <label className="flex-1">
-                    <div className="flex flex-col gap-1">
-                      <p className="text-sm font-light">Year</p>
-                      <input
-                        name="year"
-                        type="number"
-                        className="px-4 py-2 bg-card-secondary rounded-lg outline-none placeholder:text-sm [appearance:textfield] [&::-webkit-outer-spin-button]:appearance-none [&::-webkit-inner-spin-button]:appearance-none"
-                        placeholder="1448"
-                        min={0}
-                        max={new Date().getFullYear()}
-                        onChange={(e) => handleChange("year", e.target.value)}
-                      />
-                      {touchedFields.year && errors?.year && (
-                        <p className="text-red-500 text-xs">
-                          {errors.year._errors[0].toString()}
-                        </p>
-                      )}
-                    </div>
-                  </label>
-                  <label className="flex-1">
-                    <div className="flex flex-col gap-1">
-                      <p className="text-sm font-light">Language</p>
-                      <input
-                        name="language"
-                        type="text"
-                        className="px-4 py-2 bg-card-secondary rounded-lg outline-none placeholder:text-sm"
-                        placeholder="English"
-                        onChange={(e) =>
-                          handleChange("language", e.target.value)
-                        }
-                      />
-                      {touchedFields.language && errors?.language && (
-                        <p className="text-red-500 text-xs">
-                          {errors.language._errors[0].toString()}
-                        </p>
-                      )}
-                    </div>
-                  </label>
-                </div>
-                <div>
-                  <label className="flex-1">
-                    <div className="flex flex-col gap-1">
-                      <p className="text-sm font-light">Thumbnail Url</p>
-                      <input
-                        name="thumbnail"
-                        type="text"
-                        placeholder="https://example.com/thumbnail.jpg"
-                        value={thumbnailUrl}
-                        onChange={(e) => {
-                          setThumbnailUrl(e.target.value);
-                          handleChange("thumbnail", e.target.value);
-                        }}
-                        className="px-4 py-2 bg-card-secondary rounded-lg outline-none placeholder:text-sm"
-                      />
-                    </div>
-                  </label>
-                </div>
-              </div>
-              {thumbnailUrl && (
-                <div className="flex items-center h-[100%] ml-4">
-                  <img
-                    src={thumbnailUrl}
-                    alt="Thumbnail"
-                    className="max-h-[230px] min-h-[230px] object-fill border rounded"
-                    onError={(e) => {
-                      e.currentTarget.style.display = "none";
-                    }}
-                  />
-                </div>
-              )}
-            </div>
-          </div>
+          <GeneralSection
+            handleChange={handleChange}
+            errors={errors}
+            touchedFields={touchedFields}
+            setThumbnailUrl={setThumbnailUrl}
+            thumbnailUrl={thumbnailUrl}
+          />
           <div className="bg-card-secondary h-[2px] w-full rounded-lg"></div>
-          <div className="flex w-full gap-4">
-            <div className="w-1/4">
-              <h1 className="text-xl font-medium">Description</h1>
-              <h4 className="text-sm font-light">
-                Add the synopsis or a general description of the book
-              </h4>
-            </div>
-            <div className="flex-1">
-              <TextArea
-                className="resize-none w-full px-4 py-4 text-sm bg-card-secondary"
-                rows={7}
-                placeholder="Add description..."
-                name="text"
-                onChange={(e) => handleChange("description", e.target.value)}
-              />
-              {touchedFields.description && errors?.description && (
-                <p className="text-red-500 text-xs">
-                  {errors?.description._errors[0].toString()}
-                </p>
-              )}
-            </div>
-          </div>
+          <DescriptionSection
+            handleChange={handleChange}
+            errors={errors}
+            touchedFields={touchedFields}
+          />
           <div className="bg-card-secondary h-[2px] w-full rounded-lg"></div>
-          <div className="flex w-full gap-4">
-            <div className="w-1/4">
-              <h1 className="text-xl font-medium">Genres</h1>
-              <h4 className="text-sm font-light">
-                Select one or more genres for the book
-              </h4>
-            </div>
-            <div className="flex-1">
-              {pageLoaded ? (
-                <div className="flex flex-col gap-1">
-                  <p className="text-sm font-light mb-1">
-                    Genres<span className="text-red-500">*</span>
-                  </p>
-                  <Select
-                    isMulti
-                    closeMenuOnSelect={false}
-                    name="genres"
-                    options={genres}
-                    className="bg-card-secondary"
-                    classNamePrefix="select"
-                    components={animatedComponents}
-                    styles={selectStyles}
-                    onChange={(selectedOptions) => {
-                      const selectedValues = (
-                        selectedOptions as GenreOption[]
-                      ).map((option) => option.value);
-                      handleChange("genres", selectedValues);
-                    }}
-                    required
-                  />
-                  {touchedFields.genres && errors?.genres && (
-                    <p className="text-red-500 text-xs">
-                      {errors?.genres._errors[0].toString()}
-                    </p>
-                  )}
-                </div>
-              ) : (
-                <label className="flex-1">
-                  <div className="flex flex-col gap-1">
-                    <p className="text-sm font-light">
-                      Genres<span className="text-red-500">*</span>
-                    </p>
-                    <select
-                      name="genres"
-                      multiple
-                      className="h-40 px-4 py-2 bg-card-secondary rounded-lg outline-none [&::-webkit-scrollbar]:w-1 [&::-webkit-scrollbar-track]:bg-card-secondary [&::-webkit-scrollbar-thumb]:bg-gray-300 [&::-webkit-scrollbar-thumb]:rounded-lg [&::-webkit-scrollbar:horizontal]:h-1
-    [&::-webkit-scrollbar:vertical]:w-1 [&::-webkit-scrollbar-corner]:bg-transparent"
-                      required
-                      onChange={(e) => {
-                        const selectedOptions = Array.from(
-                          e.currentTarget.selectedOptions
-                        ).map((option) => option.value);
-                        handleChange("genres", selectedOptions);
-                      }}
-                    >
-                      {genres.map((genre) => (
-                        <option
-                          key={genre.value}
-                          value={genre.value}
-                          className="hover:bg-accent"
-                        >
-                          {genre.label}
-                        </option>
-                      ))}
-                    </select>
-                    <p className="text-xs text-muted-foreground">
-                      Hold Ctrl (or Cmd on Mac) to select multiple genres.
-                    </p>
-                  </div>
-                </label>
-              )}
-            </div>
-          </div>
+          <GenresSection
+            handleChange={handleChange}
+            errors={errors}
+            touchedFields={touchedFields}
+            pageLoaded={pageLoaded}
+            genres={genres}
+            selectStyles={selectStyles}
+          />
           <div className="bg-card-secondary h-[2px] w-full rounded-lg"></div>
           <div className="flex w-full gap-4">
             <div className="w-1/4">
